feat: collect open_db_connections_count if on Linux or MacOS

+++ project/techtrends/app.py
@@ -51,24 +51,27 @@ def get_open_db_connections_count():
     Returns:
         int: returns the number of current connections to the DATABASE_FILE and -1 on error
     """
-    try:
-        TIMEOUT = 20
-        p1 = subprocess.Popen(["lsof", DATABASE_FILE], stdout=subprocess.PIPE)
-        p2 = subprocess.Popen(
-            ["wc", "-l"], stdin=p1.stdout, stdout=subprocess.PIPE)
-        # extract the the 1st element of the output tuple, then convert the bytes to string then remove the new line from string then subtract 1 to ignore header line
-        db_connection_count = int(p2.communicate(timeout=TIMEOUT)[
-                                  0].decode("utf-8").replace('\n', '')) - 1
-        db_connection_count = 0 if db_connection_count == -1 else db_connection_count
-        app.logger.debug(f"db_connection_count: {db_connection_count}")
-    except subprocess.TimeoutExpired:
-        p2.kill()
-        app.logger.error(f"Error: process exceeded {TIMEOUT}")
-        db_connection_count = -1
-    except sqlite3.OperationalError as e:
-        app.logger.error(f"Error: {e}")
-        db_connection_count = -1
-    return db_connection_count
+    if platform.system() not in ['Windows', 'Java']:
+        try:
+            TIMEOUT = 20
+            p1 = subprocess.Popen(["lsof", DATABASE_FILE], stdout=subprocess.PIPE)
+            p2 = subprocess.Popen(
+                ["wc", "-l"], stdin=p1.stdout, stdout=subprocess.PIPE)
+            # extract the the 1st element of the output tuple, then convert the bytes to string then remove the new line from string then subtract 1 to ignore header line
+            db_connection_count = int(p2.communicate(timeout=TIMEOUT)[
+                                    0].decode("utf-8").replace('\n', '')) - 1
+            db_connection_count = 0 if db_connection_count == -1 else db_connection_count
+            app.logger.debug(f"db_connection_count: {db_connection_count}")
+        except subprocess.TimeoutExpired:
+            p2.kill()
+            app.logger.error(f"Error: process exceeded {TIMEOUT}")
+            db_connection_count = -1
+        except sqlite3.OperationalError as e:
+            app.logger.error(f"Error: {e}")
+            db_connection_count = -1
+        return db_connection_count
+    else:
+        return None
 
 
 def get_posts_count():
@@ -223,14 +226,15 @@ def metrics():
     res = dict()
     status_code = 200
     try:
-        # open_db_connections_count = get_open_db_connections_count()
+        open_db_connections_count = get_open_db_connections_count()
         post_count = get_posts_count()
     except Exception as e:
         status_code = 500
         res['error'] = str(e)
         app.logger.error(f"MetricsError: {e}")
     else:
-        # res['open_db_connections_count'] = open_db_connections_count
+        if open_db_connections_count:
+            res['open_db_connections_count'] = open_db_connections_count
         res['db_connection_count'] = db_conn.count
         res['post_count'] = post_count
         app.logger.debug(f"metrics: {json.dumps(res)}")
